+++ django_test_app/server/apps/model_simple/urls.py
@@ -15,4 +15,5 @@ router = Router(
             name='user_detailed',
         ),
     ],
+    tags=['model_simple'],
 )